vm: standalone emitEVMProfile
